목표인증 등록 시 이미지 미리보기, 삭제 기능을 추가하고 모달창을 닫아주는 기능 추가
